Preparing for admin levels/bitvalues.
Updated data.db to reflect the additions of XP, levels, player classes, and admin levels. 

Removed some unnecessary things in the database like player accounts that were registered using this database.

git-svn-id: http://openrp.googlecode.com/svn/trunk@91 3c299bc4-26c0-bb6f-2937-d8014291f781

+++ source/game/g_admin.cpp
@@ -39,7 +39,7 @@ M_Cmd_Status_f - MJN
 */
 void M_Cmd_Status_f( gentity_t * ent)
 {
-	int i;
+	short int i;
 	int clientNum;
 	if ( ent->client->sess.spectatorState != SPECTATOR_FOLLOW ) {
 		if (ent->client->sess.openrpIsAdmin ){
